\name{mapConsValue}
\alias{mapConsValue}
\title{Get the value of an appian constant}
\description{
Given the name of the constant, get the value for that constant.
}
\usage{
mapConsValue(x, map)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{the name of the constant either as a string or an R language object (a symbol or a call)}
  \item{map}{the data.frame computed via \code{\link{mkSummary}}}
}
\value{

}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
