\name{table2df}
\alias{table2df}
\alias{compareTypeCounts}
\alias{mkCountDfs}
\alias{mergeCounts}
\title{Utility functions for working tables of counts}
\description{
}
\usage{
table2df(x, map = NULL)
mkCountDfs(map, syms)
compareTypeCounts(...) 
}
\arguments{
  \item{x}{}
  \item{map}{}
  \item{syms}{}
  \item{\dots}{}
}
\value{
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
