\name{AppianTypesNS}
\alias{AppianTypesNS}
\title{The Appian namespace for used in XML and XPath queries}
\description{
  This is exported so it is available for anyone
  working directly with the XML content and using XPath queries.
}
\usage{AppianTypesNS}
\format{
  The format is:
 Named chr "http://www.appian.com/ae/types/2009"
 - attr(*, "names")= chr "x"
}

